added confrm delete to chapter

+++ client-admin/src/app/pages/bona-fide-list/bona-fide-list.component.ts
@@ -4,12 +4,18 @@ import {
   MatSort,
   MatPaginator,
   MatTableDataSource,
+  MatDialog,
 } from '@angular/material';
 import { AppService } from '@app/shared/app.service';
 import { FormBuilder } from '@angular/forms';
 import { Router } from '@angular/router';
 import { bonaFideservice } from '@app/shared/bonafide.service';
 import { MenuRoles } from '@app/models/enums';
+import { LanguageService } from '@app/shared/Language.service';
+import {
+  ConfirmDialogModel,
+  ConfirmDialogComponent,
+} from '@app/components/confirm-dialog/confirm-dialog.component';
 
 @Component({
   selector: 'app-bona-fide-list',
@@ -33,7 +39,9 @@ export class BonaFideListComponent implements OnInit, AfterViewInit {
     private app: AppService,
     private fb: FormBuilder,
     private bonafidesService: bonaFideservice,
-    private router: Router
+    private router: Router,
+    private languageService: LanguageService,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
@@ -77,9 +85,35 @@ export class BonaFideListComponent implements OnInit, AfterViewInit {
     }
   }
 
-  async deleteBonafide(id: string) {
-    await this.bonafidesService.delete(id);
-    this.LoadAgencies();
+  async deleteConfirm(id: string) {
+    const message = await this.languageService.translate
+      .get('CHAPTER.ARE_YOU_SURE_DELETE')
+      .toPromise();
+
+    const title = await this.languageService.translate
+      .get('CHAPTER.COMFIRMATION')
+      .toPromise();
+
+    const dialogData = new ConfirmDialogModel(
+      title,
+      message,
+      true,
+      true,
+      false
+    );
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      maxWidth: '400px',
+      data: dialogData,
+    });
+
+    dialogRef.afterClosed().subscribe(async (dialogResult) => {
+      if (dialogResult) {
+        console.log(id);
+        await this.bonafidesService.delete(id);
+        this.LoadAgencies();
+      }
+    });
   }
 
   editBonafide(id: number) {
