+++ client-admin/src/app/pages/chapter-list/chapter-list.component.ts
@@ -1,15 +1,27 @@
-import { Component, OnInit, ViewChild, Input } from '@angular/core';
+import {
+  Component,
+  OnInit,
+  ViewChild,
+  Input,
+  ChangeDetectorRef,
+} from '@angular/core';
 import {
   PageEvent,
   MatSort,
   MatPaginator,
   MatTableDataSource,
+  MatDialog,
 } from '@angular/material';
 import { AppService } from '@app/shared/app.service';
 import { FormBuilder } from '@angular/forms';
 import { AgencyService } from '@app/shared/agency.service';
 import { Router } from '@angular/router';
 import { ChapterServiceService } from '@app/shared/chapter-service.service';
+import {
+  ConfirmDialogModel,
+  ConfirmDialogComponent,
+} from '@app/components/confirm-dialog/confirm-dialog.component';
+import { LanguageService } from '@app/shared/Language.service';
 
 @Component({
   selector: 'app-chapter-list',
@@ -31,14 +43,19 @@ export class ChapterListComponent implements OnInit {
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
   @Input() id: string;
+
   constructor(
     private app: AppService,
     private fb: FormBuilder,
     private chapterService: ChapterServiceService,
-    private router: Router
+    private router: Router,
+
+    private languageService: LanguageService,
+    private dialog: MatDialog
   ) {}
 
   ngOnInit(): void {
+    console.log(this.id);
     //TODO: Acces
     //this.editAccess = this.app.checkMenuRoleAccess(MenuRoles.AGENCIES_UPDATE);
     //this.createAccess = this.app.checkMenuRoleAccess(MenuRoles.AGENCIES_CREATE);
@@ -49,10 +66,10 @@ export class ChapterListComponent implements OnInit {
   }
 
   ngAfterViewInit(): void {
-    this.LoadAgencies();
+    this.LoadChapters();
   }
 
-  private LoadAgencies() {
+  private LoadChapters() {
     this.chapterService.getByBonafideById(this.id).subscribe(
       (res) => {
         this.loading = true;
@@ -79,17 +96,46 @@ export class ChapterListComponent implements OnInit {
     }
   }
 
-  async deleteAgency(id: string) {
-    await this.chapterService.delete(id);
-    this.LoadAgencies();
+  async deleteConfirm(id: string) {
+    const message = await this.languageService.translate
+      .get('CHAPTER.ARE_YOU_SURE_DELETE')
+      .toPromise();
+
+    const title = await this.languageService.translate
+      .get('CHAPTER.COMFIRMATION')
+      .toPromise();
+
+    const dialogData = new ConfirmDialogModel(
+      title,
+      message,
+      true,
+      true,
+      false
+    );
+
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      maxWidth: '400px',
+      data: dialogData,
+    });
+
+    dialogRef.afterClosed().subscribe(async (dialogResult) => {
+      if (dialogResult) {
+        console.log(id);
+        await this.chapterService.delete(id);
+        this.LoadChapters();
+      }
+    });
+  }
+  onSubmit(task: any) {
+    throw new Error('Method not implemented.');
   }
 
-  editAgency(id: number) {
-    this.router.navigate(['/home/chapter', id]);
+  editChapter(chapterId: number) {
+    this.router.navigate(['/home/chapter', this.id, chapterId]);
   }
 
   goToNew() {
-    this.router.navigate(['/home/chapter']);
+    this.router.navigate(['/home/chapter', this.id]);
   }
 
   doFilter(value: any) {
